refactor(bench): extract the token-count measurement core into internal/bench
The corpus, the insertion protocol (fixed-seed shuffle, one-day-per-session
backdating) and the arithmetic move verbatim into package bench, shared by
two entry points: go run ./benchmarks/token_count (the documented
reproduction path) and, in the next commit, graymatter bench — so a user
without a Go toolchain can audit published numbers from the installed binary.

Zero behaviour change by construction: benchmarks/token_count/main_test.go,
which parses every published table and compares it against a live run at 2%
tolerance, is untouched and passes. main.go is now a thin wrapper whose
runBenchmark delegates; its result type aliases bench.TokenResult.

retrieval_quality stays a standalone main for now: extracting its 787-line
protocol (sliding-window baseline, adaptive mode) is a separate, riskier
surgery and gates nothing today.
